+++ src/compiler/tokens/expressions.ts
@@ -136,7 +136,7 @@ export abstract class Expression extends CompilableParseToken {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected abstract translateCtxAndChildren(): Promise<Array<any>>;
+	public abstract translateCtxAndChildren(): Promise<Array<any>>;
 
 	/**
 	 * The antlr context containing the antlr4 metadata for this expression.
@@ -185,7 +185,7 @@ export class NumberPrimaryExpression extends ConstantExpression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -194,7 +194,7 @@ export class NumberPrimaryExpression extends ConstantExpression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -239,7 +239,7 @@ export class CharacterPrimaryExpression extends ConstantExpression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -248,7 +248,7 @@ export class CharacterPrimaryExpression extends ConstantExpression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -283,7 +283,7 @@ export class ListPrimaryExpression extends ConstantExpression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -292,7 +292,7 @@ export class ListPrimaryExpression extends ConstantExpression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -335,7 +335,7 @@ export class StringPrimaryExpression extends ConstantExpression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -344,7 +344,7 @@ export class StringPrimaryExpression extends ConstantExpression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		return [`"${this.stringContent}"`];
 	}
 
@@ -386,7 +386,7 @@ export class IdentifierPrimaryExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -395,7 +395,7 @@ export class IdentifierPrimaryExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		return [this.identifierValue];
 	}
 
@@ -431,7 +431,7 @@ export class FStringPrimaryExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -440,7 +440,7 @@ export class FStringPrimaryExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -475,7 +475,7 @@ export class TangledPrimaryExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -484,11 +484,11 @@ export class TangledPrimaryExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO! Add tests for this
 		let genCode: Array<string> = [];
 		for (let child of this._children) {
-			genCode = genCode.concat(await child.compileCtx());
+			genCode = genCode.concat(await child.translateCtxAndChildren());
 		}
 		return genCode;
 	}
@@ -525,7 +525,7 @@ export class IncrementOrDecrementExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -534,7 +534,7 @@ export class IncrementOrDecrementExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -570,7 +570,7 @@ export class ArraySpecifierExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -579,7 +579,7 @@ export class ArraySpecifierExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -641,7 +641,7 @@ export class FunctionCallPostfixExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// Assert that the function exists
 		this.programCtx.assert(this).functionIsDefined(this.identifier);
 	}
@@ -651,7 +651,7 @@ export class FunctionCallPostfixExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// Get the function
 		const func = <BuiltInFunction | ScopeFunctionDeclaration>this.programCtx.getGlobalFunction(this.identifier);
 
@@ -664,7 +664,7 @@ export class FunctionCallPostfixExpression extends Expression {
 		const identifier = func instanceof ScopeFunctionDeclaration ? func.identifier : `_kipperGlobal_${func.identifier}`;
 
 		// Compile the arguments
-		const args: Array<string> = argListCtx ? await argListCtx.compileCtx() : [];
+		const args: Array<string> = argListCtx ? await argListCtx.translateCtxAndChildren() : [];
 
 		// Return the compiled function call
 		return [identifier, "(", ...args, ")"];
@@ -701,7 +701,7 @@ export class ArgumentExpressionList extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -710,10 +710,10 @@ export class ArgumentExpressionList extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		let genCode: Array<string> = [];
 		for (let child of this.children) {
-			genCode = [...genCode, ...(await child.compileCtx())];
+			genCode = [...genCode, ...(await child.translateCtxAndChildren())];
 		}
 		return genCode;
 	}
@@ -744,7 +744,7 @@ export class IncrementOrDecrementUnaryExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -753,7 +753,7 @@ export class IncrementOrDecrementUnaryExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -798,7 +798,7 @@ export class OperatorModifiedUnaryExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -807,7 +807,7 @@ export class OperatorModifiedUnaryExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -838,7 +838,7 @@ export class CastOrConvertExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -847,7 +847,7 @@ export class CastOrConvertExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -887,7 +887,7 @@ export class MultiplicativeExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -896,7 +896,7 @@ export class MultiplicativeExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -934,7 +934,7 @@ export class AdditiveExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -943,7 +943,7 @@ export class AdditiveExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -989,7 +989,7 @@ export class RelationalExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -998,7 +998,7 @@ export class RelationalExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -1038,7 +1038,7 @@ export class EqualityExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -1047,7 +1047,7 @@ export class EqualityExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -1087,7 +1087,7 @@ export class LogicalAndExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -1096,7 +1096,7 @@ export class LogicalAndExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -1136,7 +1136,7 @@ export class LogicalOrExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -1145,7 +1145,7 @@ export class LogicalOrExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -1183,7 +1183,7 @@ export class ConditionalExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -1192,7 +1192,7 @@ export class ConditionalExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -1229,7 +1229,7 @@ export class AssignmentExpression extends Expression {
 	 * Semantic analysis for the code inside this parse token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 */
-	protected async semanticAnalysis(): Promise<void> {
+	public async semanticAnalysis(): Promise<void> {
 		// TODO!
 	}
 
@@ -1238,7 +1238,7 @@ export class AssignmentExpression extends Expression {
 	 *
 	 * Every item in the array represents a token of the expression.
 	 */
-	protected async translateCtxAndChildren(): Promise<Array<string>> {
+	public async translateCtxAndChildren(): Promise<Array<string>> {
 		// TODO!
 		return [];
 	}